refactor(parser): extract AST definitions into separate module
- Created src/parser/ directory structure
- Extracted AST type definitions to src/parser/ast.rs (~400 lines)
- Moved parsing logic to src/parser/mod.rs (~2,650 lines)
- Removed original src/parser.rs (was 3,793 lines)
- Added comprehensive documentation for AST module
- All 1,224+ tests pass successfully
- Maintains backward compatibility with existing imports

This is Step 2.1 of Phase 2 parser refactoring, separating data
structures from parsing logic for better code organization.
